pyproject.toml
[project]
name = "adaptive-ui-agent"
version = "0.1.0"
description = "Visual RL Agent with VQ-VAE discrete representations for continual learning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CDA", email = "cda@example.com"},
]
keywords = ["reinforcement-learning", "vq-vae", "continual-learning", "vision"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pygame>=2.5.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0.0",
    "tensorboard>=2.14.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.scripts]
train = "scripts.run_training:main"
demo = "scripts.demo:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "env*", "vision*", "interaction*", "configs*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
